meta-rz-bsp: Add bootparameter recipe
The bootrom on the RZ/G2L family of SoCs expects the intial bootloader to
have a special header containing the aligned filesize of the binary and a
magic number.

This bootparameter application will take an input file and create the
required header.
The recipe will then append the header onto the orignal binary, ready to
flash to the platform.

Signed-off-by: Biju Das <biju.das.jz@bp.renesas.com>
Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>