meta-rz-bsp: u-boot-renesas.inc: Make sure package u-boot is provided
u-boot.inc provided by poky PROVIDES "virtual/bootloader" and not
"u-boot".

This can cause us an issue when a different recipe DEPENDS on "u-boot"
as we may be building with our u-boot-renesas recipe and bitbake will
fail to find anything that provides "u-boot".

By specifically adding the "u-boot" PROVIDES here we can work around
this issue.

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>