meta-rz-bsp: trusted-firmware-a: Add support for smarc-rzg2l
The smarc-rzg2l platform expects the bootloader files in a specific
format. This is done using the bootparameter and fip tools. TF-A
already includes support for creating fip packages, so make use of it.

Mainline trusted-firmware-a doesn't support smarc-rzg2l, so use
COMPATIBLE_MACHINE to prevent use of trusted-firmware-a_2.9.0.bb.

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>