gitlab-ci: Remove git2 dir from DL_DIR before caching
When bitbake makes use of DL_DIR it will use a repository found in the
${DL_DIR}/git2 directory before looking elsewhere for it.

We've seen occasional issues when using this approach with GitLab CI
caching. We are already generating the mirror tarballs
(BB_GENERATE_MIRROR_TARBALLS) so let's force bitbake to use them by
removing the git2 directory.

Do the removal before and after the build so we can reuse the current
DL_DIR cache and then save the "improved" copy afterwards.

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>