scripts: submit-lava-job.sh: Improve binary location when using GitLab CI artifacts
With the old approach we end up doing some detective work trying to work
out if a file is a symlink or not, as GitLab CI artifacts do not handle
symlinks properly.

This commit introduces a better method. It relies on the build jobs
copying the binaries we need to a known location (instead of using the
tmp/deploy/images/<machine> directory normally used).

This means that we know the files present will be the ones we want to use
and we don't need to do any other magic.

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>