meta-rz-bsp: trusted-firmware-a-renesas.inc: Use virtual/trusted-firmware-a
Because there are two recipes that both PROVIDE trusted-firmware-a, if we
switch between them we get complaints that files are being overwritten
during do_deploy:

ERROR: trusted-firmware-a-renesas-2.9.0-r0 do_deploy: Recipe trusted-firmware-a-renesas is trying to install files into a shared area when those files already exist. The files and the manifests listing them are:
  /build/tmp/deploy/images/hihope-rzg2h/bootparam_sa0.srec
    (matched in manifest-hihope_rzg2h-trusted-firmware-a.deploy)
  /build/tmp/deploy/images/hihope-rzg2h/cert_header_sa6.srec
    (matched in manifest-hihope_rzg2h-trusted-firmware-a.deploy)
  /build/tmp/deploy/images/hihope-rzg2h/bl2-hihope-rzg2h.srec
    (matched in manifest-hihope_rzg2h-trusted-firmware-a.deploy)
  /build/tmp/deploy/images/hihope-rzg2h/bl31-hihope-rzg2h.srec
    (matched in manifest-hihope_rzg2h-trusted-firmware-a.deploy)
Please adjust the recipes so only one recipe provides a given file.

If we set PROVIDES="virtual/trusted-firmware-a", bitbake will remove the
files in question before they are overwritten as it is expecting this use
case.

Also update the files that reference trusted-firmware-a to use the virtual
identifier.

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>