gitlab-ci: Add ability to clear the caches as part of the pipeline
It's good to flush the GitLab CI runner caches now and then, but the only
control that GitLab provides is manaul and it destroys every cache in the
project.

The changes introduced in this commit allow the manual removal of
individual caches. They can only be removed by passing variables into the
CI pipeline. For example, a schedule could be created for the master
branch that sets CLEAR_SSTATE_CACHE="true". This would delete the sstate
cache used by the master branch alone.

Other options are CLEAR_DEPENDENCIES_CACHE and CLEAR_DOWNLOADS_CACHE.

Note: We need to stop jobs running if the pipeline source is
"merge_request_event" to avoid duplicate pipelines appearing in merge
requests.

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>