bootparameter: add missing gap of 6 sectors
When writing bl2_bp.bin to the SD card, nothing happens at power on.

According to:
    https://renesas.info/w/images/2/28/SD_card.png

There must be a 6 sectors gap between the boot parameters and the start of
BL2, and BL2 must start at sector 8. Currently, BL2 starts at sector 2.

Add the required gap of 6 sectors of 512 bytes each so that the boot process
can work.