gitlab-ci: Stop using GitLab CI cache to save Yocto sstate/downloads
From experience it seems that GitLab CI cache isn't great for the large
data we save when caching Yocto SSTATE_DIR and DL_DIR. We keep seeing an
array of issues caused by the way it works.

Instead, use a new NFS mount to save the sstate and downloads, using
Yocto's built-in functionality to manage them.

As part of this, move the BB_GENERATE_MIRROR_TARBALLS enablement to
kas/base.yml and remove kas/opt/gitlab-ci-cache.yml.

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>