meta-rz-bsp: machine: tf-a: Enable INIT_UNUSED_NS_EL2 flag
Since TF-A v2.10.0 commit
183329a5847d ("refactor(cm): introduce INIT_UNUSED_NS_EL2 macro")
introduced a breaking change that requires platforms that are without
NS-EL2 to enable the INIT_UNUSED_NS_EL2 flag.

Without this flag we fail to boot into BL31.

Enable this flag in the relevant machine configurations to keep all the
related TF-A build flags in one place.

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>