meta-rz-bsp: Add new renesas-image-demo image
The renesas-image-demo.bb recipe provides a more feature rich image that
can be used to showcase the functionality of Renesas RZ SoCs.

It includes Weston/Wayland, Panfrost and a number of packages useful for
hardware debugging such can-utils and i2c-tools.

If not building with Kas/Kas menu the following will need to be added to
local.conf:
  DISTRO_FEATURES += " opengl wayland"

The image is dependent on the meta-rz-panfrost layer.

Because this image includes Panfrost support, it is currently only
supported for the smarc-rzg2l and smarc-rzg2lc machines. We'll make
Panfrost optional and add support for other machines in the future.

Signed-off-by: Sachin Dominic <sachin.dominic.zn@renesas.com>
Signed-off-by: Paul Barker <paul.barker.ct@bp.renesas.com>
Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>