Add pre-commit configuration
This allows us to use the pre-commit tool [1] to check for common
mistakes and to ensure that relevant best practices are followed.

For PyMarkdown, rules MD013 (line length), MD033 (inline HTML) and MD034
(bare URLs) are ignored as they are too difficult to comply with while
achieving the readme layout we want.

[1]: https://pre-commit.com/

Signed-off-by: Paul Barker <paul.barker.ct@bp.renesas.com>