gitlab-ci: Add deploy job to automatically push updates to kirkstone branch
We've changed the structure of the repository so that we have a branch for
each version of Yocto we support. This is in line with standard
OpenEmbedded practices.

For now, both of these branches are identical.

To save manual work, add a CI job that automates the process of keeping
both branches in sync.
Whenever a pipeline is run on the default branch (scarthgap) it will
automatically push the changes to the kirkstone branch.

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>