meta-rz-bsp: trusted-firmware-a: Fix fiptool linkage
In meta-arm, the trusted-firmware-a.inc file handles patching the
fiptool Makefile to include the correct linker arguments for native
compilation under Yocto. However, in TF-A 2.10.0, the variable
containing linker arguments in the fiptool Makefile was changed from
`LDLIBS` to `LDOPTS` and this complicates things a bit as we support
multiple Yocto branches and multiple TF-A versions.

In scarthgap, meta-arm uses TF-A 2.10.x and so handles the LDOPTS
variable. So to build TF-A 2.9.0 in this case we need to handle the
LDLIBS variable ourselves.

In kirkstone, meta-arm uses TF-A 2.6 or 2.7 and so handles the LDLIBS
variable. So to build TF-A 2.10.8 in this case we need to handle the
LDOPTS variable ourselves.

Prior to this patch, after trying to build TF-A 2.9.0 with the scarthgap
branch, we see that no search paths have been embedded in the fiptool
binary:

    $ ldd git/tools/fiptool/fiptool
        linux-vdso.so.1 (0x00007fff8b0b7000)
        libcrypto.so.3 => not found
        libc.so.6 => /lib/x86_64-linux-gnu/libc.so.6 (0x00007f7cd9617000)
        /lib64/ld-linux-x86-64.so.2 (0x00007f7cd981f000)

After this patch, with paths shortened for clarity:

    $ ldd git/tools/fiptool/fiptool
        linux-vdso.so.1 (0x00007ffd705e8000)
        libcrypto.so.3 => [...]/tmp/work/smarc_rzv2l-poky-linux/trusted-firmware-a-renesas/2.9.0/recipe-sysroot-native/usr/lib/libcrypto.so.3 (0x00007f7d22000000)
        libpthread.so.0 => /lib/x86_64-linux-gnu/libpthread.so.0 (0x00007f7d22615000)
        libc.so.6 => /lib/x86_64-linux-gnu/libc.so.6 (0x00007f7d21e0e000)
        libdl.so.2 => /lib/x86_64-linux-gnu/libdl.so.2 (0x00007f7d2260f000)
        [...]/tmp/sysroots-uninative/x86_64-linux/lib/ld-linux-x86-64.so.2 => /lib64/ld-linux-x86-64.so.2 (0x00007f7d2264e000)

Fixes #7.

Signed-off-by: Paul Barker <paul.barker.ct@bp.renesas.com>