meta-rz-bsp: bootparameter: Add handling for new do_unpack target
Yocto Styhead made changes so that files from do_unpack now unpack to
$WORKDIR/sources-unpack/ rather than $WORKDIR/.
We now need to set $UNPACKDIR and $S to deal with this.
This in turn breaks do_unpack when using older versions of Yocto, so we
need to add some hackery in a custom do_unpack_extra function to make
all Yocto versions behave the same.

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>