gitlab-ci: Push to GitHub mirror after default branch build
Currently we have a push mirror from GitLab to GitHub.
This can lead to branches being out of sync as we only push to the
non-default release branches after a CI run has completed.
For example, if a CI run on the default (scarthgap) branch fails then the
GitHub mirror will end up containing the new (maybe broken) code for that
branch but not the others.

For the new behaviour we will schedule routine production runs to test the
default branch and then if successful we will push to the GitHub mirror.

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>